# docs/swagger.yaml
definitions:
  dto.CreateKeywordRequest:
    properties:
      site_id:
        type: integer
      value:
        type: string
    required:
    - site_id
    - value
    type: object
  dto.CreateSiteRequest:
    properties:
      domain:
        type: string
      name:
        type: string
    required:
    - domain
    - name
    type: object
  dto.DeleteKeywordResponse:
    properties:
      message:
        type: string
    type: object
  dto.DeleteSiteResponse:
    properties:
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  dto.KeywordResponse:
    properties:
      id:
        type: integer
      site_id:
        type: integer
      value:
        type: string
    type: object
  dto.SiteResponse:
    properties:
      domain:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/keywords:
    get:
      description: Get list of keywords for a specific site
      parameters:
      - description: Site ID
        in: query
        name: site_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.KeywordResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get keywords by site
      tags:
      - keywords
    post:
      consumes:
      - application/json
      description: Create a new keyword for tracking
      parameters:
      - description: Keyword data
        in: body
        name: keyword
        required: true
        schema:
          $ref: '#/definitions/dto.CreateKeywordRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.KeywordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new keyword
      tags:
      - keywords
  /api/keywords/{id}:
    delete:
      description: Delete a keyword and all its tracking data
      parameters:
      - description: Keyword ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteKeywordResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a keyword
      tags:
      - keywords
  /api/sites:
    get:
      description: Get list of all tracked sites
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.SiteResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all sites
      tags:
      - sites
    post:
      consumes:
      - application/json
      description: Create a new site for tracking
      parameters:
      - description: Site data
        in: body
        name: site
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSiteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SiteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new site
      tags:
      - sites
  /api/sites/{id}:
    delete:
      description: Delete a site and all its tracking data
      parameters:
      - description: Site ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteSiteResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete a site
      tags:
      - sites
swagger: "2.0"